fix cupy fft dtype

+++ pyscf/pbc/tools/pbc.py
@@ -105,12 +105,12 @@ elif FFT_ENGINE == 'CUPY':
     def _fftn_wrapper(a):
         a = lib.device_put(a)
         a_fft = cupy.fft.fftn(a, axes=(1,2,3))
-        a_fft = lib.device_get(a_fft)
+        a_fft = lib.device_get(a_fft, dtype=np.complex128)
         return a_fft
     def _ifftn_wrapper(a):
         a = lib.device_put(a)
         a_ifft = cupy.fft.ifftn(a, axes=(1,2,3))
-        a_ifft = lib.device_get(a_ifft)
+        a_ifft = lib.device_get(a_ifft, dtype=np.complex128)
         return a_ifft
 
 #?elif:  # 'FFTW+BLAS'
